% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformMatrixCol.R
\name{transformMatrixCol}
\alias{transformMatrixCol}
\title{Tidy assay data column names}
\usage{
transformMatrixCol(expMat, Sep, nSep, pos)
}
\arguments{
\item{expMat}{A matrix data frame.}

\item{Sep}{The delimiter that joins the appendage to desired column name}

\item{nSep}{The count of the delimiter in the each column name.}

\item{pos}{The position of the desired column name in Assay column name}
}
\value{
A same matrix data frame, with desired column name.
}
\description{
In transcriptomics study, one of the objectives is to compare gene
expression between two or more groups in other to explore the molecular
mechanism that differentiates the groups. However, before performing differential
expression genes analysis on expression data, it is important to ensure that
the column names of the assay data is same with sample ID of the clinical or metadata.
Unfortunately, most count data obtained from raw unprocessed data have their column
names in a messy form.\cr

This function therefore help to transform the assay data column names to suit
the sample name in the metadata. Thereby handling the daunting task of removing the attached
appendages and saving time to focus on what matters.
}
\details{
Data wrangling to ease differential gene expression workflow
}
\examples{

\dontrun{
library(tidyAssayData)
transformMatrixCol(Exprdata, Sep = ".", nSep = 2, pos = 2)
}


}
