% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformMatrixCol.R
\name{transformMatrixCol}
\alias{transformMatrixCol}
\title{Tidy assay data column names}
\usage{
transformMatrixCol(expMat, Sep, nSep, pos)
}
\arguments{
\item{expMat}{A matrix data frame.}

\item{Sep}{The delimiter that joins the appendage to desired column name}

\item{nSep}{The count of the delimiter in the each column name.}

\item{pos}{The position of the desired column name in Assay column name}
}
\value{
a data frame, with desired column name.
}
\description{
Before performing differential expression genes analysis on expression data, it is important
to ensure that the column names of the assay data is same with sample ID of the clinical or metadata.
This function therefore help to transform the assay data column names to suit the sample name in the metadata.
}
\details{
Title
}
\examples{
transformMatrixCol(Exprdata, Sep = ".", nSep = 2, pos = 2)
}
